refactor: AndroidEntryPoints Annotation 사용으로 변경 - #6
