+++ app/src/main/java/com/beomjo/whitenoise/ui/common/ProgressDialogFragment.kt
@@ -10,19 +10,15 @@ import com.beomjo.whitenoise.databinding.FragmentProgressDialogBinding
 import com.beomjo.whitenoise.di.common.CommonComponent
 import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
+import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.EntryPointAccessors
 import dagger.hilt.components.SingletonComponent
 
+@AndroidEntryPoint
 class ProgressDialogFragment : BaseDialogFragment<FragmentProgressDialogBinding>(
     R.layout.fragment_progress_dialog,
     ProgressDialogViewModel::class
 ) {
-
-    @InstallIn(SingletonComponent::class)
-    @EntryPoint
-    interface ProgressDialogEntryPoints {
-        fun commonComponent(): CommonComponent.Factory
-    }
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -32,16 +28,6 @@ class ProgressDialogFragment : BaseDialogFragment<FragmentProgressDialogBinding>
         isCancelable = false
         this.dialog?.window?.setDimAmount(0.0f)
         return super.onCreateView(inflater, container, savedInstanceState)
-    }
-
-    override fun inject() {
-        activity?.application?.let {
-            val entryPoint = EntryPointAccessors.fromApplication(
-                it,
-                ProgressDialogEntryPoints::class.java
-            )
-            entryPoint.commonComponent().create()
-        } ?: kotlin.run { dismiss() }
     }
 
     companion object {
