+++ app/src/main/java/com/beomjo/whitenoise/ui/main/home/HomeFragment.kt
@@ -18,31 +18,19 @@ import com.beomjo.whitenoise.ui.main.setting.SettingFragment
 import com.beomjo.whitenoise.ui.main.track.TrackListFragment
 import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
+import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.EntryPointAccessors
 import dagger.hilt.components.SingletonComponent
 
+@AndroidEntryPoint
 class HomeFragment : BaseFragment<FragmentHomeBinding>(
     R.layout.fragment_home,
     HomeViewModel::class,
 ) {
-    @InstallIn(SingletonComponent::class)
-    @EntryPoint
-    interface HomeEntryPoints {
-        fun homeComponent(): HomeComponent.Factory
-    }
-
     private val homeViewModel: HomeViewModel by getViewModel()
 
     override val viewModelProvideOwner: ViewModelStoreOwner
         get() = activity as ViewModelStoreOwner
-
-    override fun inject() {
-        val entryPoint = EntryPointAccessors.fromApplication(
-            this.context?.applicationContext,
-            HomeEntryPoints::class.java
-        )
-        entryPoint.homeComponent().create()
-    }
 
     override fun onStart() {
         super.onStart()
