+++ app/src/main/java/com/beomjo/whitenoise/ui/main/setting/SettingFragment.kt
@@ -15,13 +15,12 @@ import com.beomjo.whitenoise.base.BaseFragment
 import com.beomjo.whitenoise.databinding.FragmentSettingBinding
 import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class SettingFragment : BaseFragment<FragmentSettingBinding>(
     R.layout.fragment_setting
 ) {
     override val viewModelProvideOwner: ViewModelStoreOwner
         get() = activity as ViewModelStoreOwner
-
-    override fun inject() {}
 
     override fun onCreateView(
         inflater: LayoutInflater,
